+++ src/buoy_detection/camera_calibration.py
@@ -111,8 +111,95 @@ def calibrate(camera_number, camera_width, camera_height, image_path):
     new_camera_matrix, roi = cv2.getOptimalNewCameraMatrix(mtx, dist, (w, h), 1, (w, h))
     return (new_camera_matrix, roi, mtx, objpoints, imgpoints)
 
-#Example usage
-#def main():
+
+
+def stereo_calibrate(camera_number1, camera_width1, camera_height1, camera_number2, camera_width2, camera_height2):
+
+
+
+    stereo_calibration_criteria = (cv2.TERM_CRITERIA_MAX_ITER + cv2.TERM_CRITERIA_EPS, 100, 1e-5)
+    stereo_calibration_flags = cv2.CALIB_FIX_INTRINSIC
+
+
+    image_path1 = generate_images(camera_number1, camera_width1, camera_height1)
+    (left_distortion_coeff, left_roi, left_mtx, left_objpoints, left_imagepoints) = calibrate(
+        camera_number1,
+        camera_width1,
+        camera_height1,
+        image_path1)
+
+    image_path2 = generate_images(camera_number2, camera_width2, camera_height2)
+    (right_distortion_coeff, right_roi, right_mtx, right_objpoints, right_imagepoints) = calibrate(
+        camera_number2,
+        camera_width2,
+        camera_height2,
+        image_path2)
+
+
+    stereocalibration_retval, left_mtx, left_distortion_coeff, \
+    right_mtx, right_distortion_coeff, R, T, E, F = cv2.stereoCalibrate(
+        left_objpoints,
+        left_imagepoints,
+        right_imagepoints,
+        left_mtx,
+        left_distortion_coeff,
+        right_mtx,
+        right_distortion_coeff,
+        (camera_height1, camera_width1),
+        criteria=stereo_calibration_criteria,
+        flags=stereo_calibration_flags)
+
+
+    (leftRectification, rightRectification, leftProjection,
+     rightProjection,
+     disparity_to_depth_map,
+     left_roi, right_roi) = cv2.stereoRectify(
+        left_mtx,
+        left_distortion_coeff,
+        right_mtx,
+        right_distortion_coeff,
+        (camera_height1, camera_width1),
+        R,
+        T, alpha=1)
+
+    left_xmap, left_ymap = cv2.initUndistortRectifyMap(
+        left_mtx,
+        left_distortion_coeff,
+        leftRectification,
+        leftProjection,
+        (camera_height1, camera_width1),
+        cv2.CV_32FC1)
+
+    right_xmap, right_ymap = cv2.initUndistortRectifyMap(
+        right_mtx,
+        right_distortion_coeff,
+        rightRectification,
+        rightProjection,
+        (camera_height1, camera_width1),
+        cv2.CV_32FC1)
+
+    return (camera_height1, camera_width1, left_xmap, left_ymap, left_roi, right_xmap, right_ymap, right_roi)
+
+
+def stereo_test():
+    (camera_height1, camera_width1, left_xmap, left_ymap, left_roi, right_xmap, right_ymap, right_roi) = stereo_calibrate(0, 640, 480, 0, 640, 480)
+    np.savez_compressed("./x", camera_height1, camera_width1, left_xmap, left_ymap, left_roi, right_xmap, right_ymap,
+                        right_roi)
+    cal = np.load("./x.npz", allow_pickle=False)
+    print(cal)
+    # image_size = tuple(calibration["image_size"])
+    # leftMapX = calibration["leftMapX"]
+    # leftMapY = calibration["leftMapY"]
+    # leftROI = tuple(calibration["leftROI"])
+    # rightMapX = calibration["rightMapX"]
+    # rightMapY = calibration["rightMapY"]
+    # rightROI = tuple(calibration["rightROI"]
+
+
+
+#def mono_test():
 #    image_path = generate_images(0, 1920, 1080)
 #    (new_mtx, roi, mtx, None, None) = calibrate(0, 1920, 1080, image_path)
-#main()
+
+
+stereo_test()
